reorganizing code and closing game loop
