+++ go.mod
@@ -1,2 +1,3 @@
 module casino
+
 go 1.22
