+++ services/blackjack/printer.go
@@ -1,4 +1,4 @@
-package main
+package blackjack
 
 import "fmt"
