+++ services/blackjack/queue.go
@@ -1,4 +1,4 @@
-package main
+package blackjack
 
 import (
 	"casino/libs/store"
