+++ services/blackjack/state.go
@@ -1,0 +1,24 @@
+package blackjack
+
+import "casino/libs/fsm"
+
+type Outcome string
+
+const (
+	Win  Outcome = "WIN"
+	Loss Outcome = "LOSS"
+	Push Outcome = "PUSH"
+)
+
+const (
+	StateTableOpen     fsm.State = "TableOpen"
+	StateTableClosed   fsm.State = "TableClosed"
+	StateShuffleCards  fsm.State = "ShuffleCards"
+	StateDealCards     fsm.State = "DealCards"
+	StateBetsOpen      fsm.State = "BetsOpen"
+	StateBetsClosed    fsm.State = "BetsClosed"
+	StateBetsSettle    fsm.State = "BetsSettle"
+	StateInsuranceTurn fsm.State = "InsuranceTurn"
+	StatePlayerTurn    fsm.State = "PlayerTurn"
+	StateDealerTurn    fsm.State = "DealerTurn"
+)
